Create marketplace layout
