+++ apps/festival/festival/src/app/marketplace/marketplace.component.scss
@@ -2,23 +2,3 @@
   display: block;
   height: 100%;
 }
-
-mat-sidenav-container {
-  height: 100%;
-}
-
-mat-sidenav-content {
-  padding-top: 80px;
-}
-mat-toolbar {
-  // position: sticky;
-  // top: 0;
-  // z-index: 2;
-
-  .logo {
-    height: 48px;
-    img {
-      height: 100%;
-    }
-  }
-}
