+++ libs/ui/src/lib/layout/marketplace/marketplace.component.html
@@ -1,0 +1,56 @@
+<mat-sidenav-container>
+  <mat-sidenav id="dashboard-navbar">
+    <ng-content></ng-content>
+  </mat-sidenav>
+
+  <mat-sidenav-content appNavContainer [appNav]="nav">
+    <!-- TOOLBAR -->
+    <app-nav #nav>
+      <mat-toolbar fxLayout="row" fxLayoutAlign="space-between center">
+        <div fxLayout="row" fxLayoutAlign="default center" fxLayoutGap="24px">
+          <button test-id="menu" mat-icon-button (click)="sidenav.toggle()">
+            <mat-icon svgIcon="menu"></mat-icon>
+          </button>
+          <a class="logo" routerLink=".">
+            <img asset="festival.svg" type="logo" alt="Archipel Content Logo" />
+          </a>
+        </div>
+        <div fxLayout="row" fxLayoutGap="24px">
+          <a test-id="heart-icon" mat-icon-button routerLink="wishlist">
+            <mat-icon
+              svgIcon="heart_fill" color="primary"
+              [matBadge]="wishlistCount$ | async" [matBadgeHidden]="!(wishlistCount$ | async)" matBadgeColor="accent" matBadgeSize="small">
+            </mat-icon>
+          </a>
+          <mat-divider vertical></mat-divider>
+          <!-- Invitations -->
+          <a mat-icon-button routerLink="invitations">
+            <ng-container *ngIf="invitationCount$ | async as count; else noInvitation">
+              <mat-icon svgIcon="send" [matBadge]="count" matBadgeColor="accent" matBadgeSize="small"></mat-icon>
+            </ng-container>
+            <ng-template #noInvitation>
+              <mat-icon svgIcon="send"></mat-icon>
+            </ng-template>
+          </a>
+          <!-- Notifications -->
+          <a mat-icon-button routerLink="notifications">
+            <ng-container *ngIf="notificationCount$ | async as count; else noNotification">
+              <mat-icon svgIcon="bell" [matBadge]="count" matBadgeColor="accent" matBadgeSize="small"></mat-icon>
+            </ng-container>
+            <ng-template #noNotification>
+              <mat-icon svgIcon="bell"></mat-icon>
+            </ng-template>
+          </a>
+          <!-- <overlay-notification-widget page="notifications"></overlay-notification-widget> -->
+          <auth-widget></auth-widget>
+        </div>
+      </mat-toolbar>
+    </app-nav>
+
+    <!-- CONTENT -->
+    <main [@routeAnimation]="prepareRoute()">
+      <router-outlet></router-outlet>
+    </main>
+
+  </mat-sidenav-content>
+</mat-sidenav-container>
