+++ libs/ui/src/lib/layout/marketplace/marketplace.component.scss
@@ -1,0 +1,24 @@
+:host {
+  display: block;
+  height: 100%;
+}
+
+mat-sidenav-container {
+  height: 100%;
+}
+
+mat-sidenav-content {
+  padding-top: 80px;
+}
+mat-toolbar {
+  // position: sticky;
+  // top: 0;
+  // z-index: 2;
+
+  .logo {
+    height: 48px;
+    img {
+      height: 100%;
+    }
+  }
+}
